Rearranged static compression
Moved raven-js to the head of the page, without compression, so that this will be applied before anything else. Then any other javascript is applied in the body. since jquery is only required for the images popup pages, that dependency has been moved to `base/photos.html`. Also all css and javascript is now compressed, removing the unnecessary `js_no_compress` and `css_no_compress` sections