Moved Magnific Popup to Bower
Before I had downloaded a customized version of magnific popup:
http://bit.ly/magnific-popup#build=inline+image+gallery+imagezoom+fastclick
So moving to Bower, I am just using the default full version for now.
In the future I want to build my own customized version, but that will
have to wait until there is better JS build support.

Also I will continue to use my own settings.scss file, but it
has to be in a subfolder so that the LESS command line can load
the folder and use it as an import for main.scss magnific-popup file